Changed the gemset name separator from / to @
This change makes a backwards incompatible change to the way gemsets are
defined and used.  You should now use this syntax:

    rvm_gemset {
      'ruby-1.9.2-p290@mygemset':
        ensure => present,
        require => Rvm_system_ruby['ruby-1.9.2-p290'];
    }

    rvm_gem {
      'ruby-1.9.2-p290@mygemset/bundler':
        ensure => present,
        require => Rvm_gemset['ruby-1.9.2-p290@mygemset'];
    }

Ruby versions are now separated from gemsets with an `@` instead of a
`/`.  Gems still use a `/` in their name to separate them from the ruby
version.  This should make it defining and using gemsets a bit cleaner
and more like rvm.

Sorry for the inconvenience.  Fixes #17.