Include ruby- prefix in hiera example.
I'm trying to get RVM with passenger working; and I started with the heira syntax.  What I've found is that the passenger support requires (1) you to be consistent with either using or omitting the ruby- prefix in order to get started, and (2) ultimately you need to include the ruby- prefix as that's what's included in the generated Apache configuration.  To save others a bit of time, I suggest updating the docs to include the prefix in the hiera example.